+++ src/main/java/subway/application/LineStationService.java
@@ -7,6 +7,7 @@ import subway.dao.LineStationDao;
 import subway.dao.StationDao;
 import subway.domain.*;
 import subway.dto.*;
+import subway.dto.response.*;
 import subway.exceptions.customexceptions.InvalidDataException;
 
 import java.util.ArrayList;
