+++ src/main/java/subway/dto/request/LineStationRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 public class LineStationRequest {
     private String upBoundStationName;
