+++ src/main/java/subway/dto/response/LineStationResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 public class LineStationResponse {
     private Long id;
