+++ src/main/java/subway/dto/response/StationsResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 import subway.domain.Station;
