+++ src/main/java/subway/exceptions/GlobalExceptionHandler.java
@@ -3,7 +3,6 @@ package subway.exceptions;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import subway.exceptions.customexceptions.InvalidDataException;
 import subway.exceptions.customexceptions.NotFoundException;
@@ -20,5 +19,10 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(NotFoundException.class)
     ResponseEntity<ExceptionResponse> handleNotFoundException(NotFoundException e) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ExceptionResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler(Exception.class)
+    ResponseEntity<ExceptionResponse> handleUnexpectedException(Exception e) {
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ExceptionResponse(DEFAULT_EXCEPTION_MESSAGE));
     }
 }
