+++ src/main/java/subway/exceptions/customexceptions/InvalidDataException.java
@@ -1,7 +1,6 @@
 package subway.exceptions.customexceptions;
 
 public class InvalidDataException extends RuntimeException {
-
     public InvalidDataException(String message) {
         super(message);
     }
